Enable curation tasks to be queued when an item's embargo is set/lifted
There is no standard DSpace event that's sent when an item's embargo status is
changed. However, we need to be able to do extra actions at this stage. This
uses the same mechanism to queue curation tasks that I'm using the queue
curation tasks when an event is received (consequently, there's been some
refactoring to rationalise the code directly connected to curation tasks).

[Task #1264: Waikato, Otago -- port media type annotator to DSpace 1.7]